+++ R/init_db_userAsmb.R
@@ -630,16 +630,6 @@ new_db_userAsmb <- function(
     );"
   )
 
-  DBI::dbExecute(
-    con,
-    "CREATE TABLE gap_evidence (
-      ID TEXT NOT NULL,
-      genus_match TEXT,
-      time_stamp INTEGER,
-      PRIMARY KEY (ID)
-    );"
-  )
-
   # Add Annotate table ----
   DBI::dbExecute(
     con,